[FIX] im_livechat: create one channel when posting multiple messages
Before this commit, sending multiple messages before the channel
creation can result in multiple channels being created.

It occurs because the post function is overriden to first persist
the channel. When the persist call is still in progress, we shouldn't
issue a new one.

task-4756758

X-original-commit: 41ca679

+++ addons/im_livechat/static/src/embed/common/livechat_service.js
@@ -68,31 +68,43 @@ export class LivechatService {
         if (!thread.isTransient) {
             return thread;
         }
-        const temporaryThread = thread;
-        const deleteTemporary = async () => {
-            await this.store.chatHub.initPromise;
-            temporaryThread.channel.chatWindow?.close({ force: true });
-            temporaryThread?.delete();
-        };
-        const savedChannel = await this._createChannel({ originThread: thread, persist: true });
-        if (!savedChannel) {
-            await deleteTemporary();
-            return;
+        if (this._persistResolvers) {
+            return this._persistResolvers.promise;
         }
-        savedChannel.fetchNewMessages();
-        this.env.services["mail.store"].initialize();
-        savedChannel.readyToSwapDeferred.then(async () => {
-            if (!savedChannel?.exists()) {
+        this._persistResolvers = Promise.withResolvers();
+        try {
+            const temporaryThread = thread;
+            const deleteTemporary = async () => {
+                await this.store.chatHub.initPromise;
+                temporaryThread.channel.chatWindow?.close({ force: true });
+                temporaryThread?.delete();
+            };
+            const savedChannel = await this._createChannel({ originThread: thread, persist: true });
+            if (!savedChannel) {
+                await deleteTemporary();
                 return;
             }
-            // Do not load unread messaes: new messages were loaded to avoid
-            // flickering, we do not want another load that would result in the
-            // same issue.
-            savedChannel.scrollUnread = false;
-            deleteTemporary();
-            savedChannel.openChatWindow({ focus: true });
-        });
-        return savedChannel;
+            savedChannel.fetchNewMessages();
+            this.env.services["mail.store"].initialize();
+            savedChannel.readyToSwapDeferred.then(async () => {
+                if (!savedChannel?.exists()) {
+                    return;
+                }
+                // Do not load unread messages: new messages were loaded to avoid
+                // flickering, we do not want another load that would result in the
+                // same issue.
+                savedChannel.scrollUnread = false;
+                deleteTemporary();
+                savedChannel.openChatWindow({ focus: true });
+            });
+            this._persistResolvers.resolve(savedChannel);
+            return savedChannel;
+        } catch (error) {
+            this._persistResolvers.reject(error);
+            throw error;
+        } finally {
+            this._persistResolvers = null;
+        }
     }
 
     /**
